removed export default, reduce tslib dependency, support expo apollo
