+++ packages/apollo-cache/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@wora/apollo-cache",
-  "version": "2.1.0",
+  "version": "2.2.0",
   "keywords": [
     "cache",
     "wora",
