+++ packages/apollo-offline/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@wora/apollo-offline",
-  "version": "2.2.0",
+  "version": "2.3.0",
   "keywords": [
     "wora",
     "cache-persist",
