+++ packages/cache-persist/__tests__/Cache-test.ts
@@ -1,9 +1,9 @@
-import { default as Cache } from '../src/Cache';
+import { Cache } from '../src';
 import { ICache } from '../src/CacheTypes';
-import createStorage from '../src/createStorage';
-import prefixLayer from '../src/layers/prefixLayer';
+import { createStorage } from '../src/createStorage';
+import { prefixLayer } from '../src/layers/prefixLayer';
 
-jest.mock('../src/createStorage', () => require.requireActual('../__mocks__/createStorage').default);
+jest.mock('../src/createStorage', () => require.requireActual('../__mocks__/createStorage'));
 
 const INITIAL_STATE = { restore: true, data: 1 };
 function sleep(ms) {
