+++ packages/cache-persist/package.json
@@ -33,12 +33,12 @@
     "test": "jest --coverage"
   },
   "dependencies": {
-    "idb": "^4.0.0",
-    "tslib": "^1.11.1"
+    "idb": "^4.0.0"
   },
   "devDependencies": {
     "rimraf": "2.6.3",
     "typescript": "3.5.2",
-    "@react-native-community/async-storage": "^1.6.1"
+    "@react-native-community/async-storage": "^1.6.1",
+    "tslib": "^1.11.1"
   }
 }
