+++ packages/cache-persist/src/index.ts
@@ -1,11 +1,11 @@
-import Cache from './Cache';
-export * from './Cache';
+import { Cache } from './Cache';
 
 export * from './CacheTypes';
 
-export { default as mutateKeysLayer } from './layers/mutateKeys';
-export { default as mutateValuesLayer } from './layers/mutateValues';
-export { default as prefixLayer } from './layers/prefixLayer';
-export { default as compose } from './utils/compose';
+export { mutateKeys as mutateKeysLayer } from './layers/mutateKeys';
+export { mutateValues as mutateValuesLayer } from './layers/mutateValues';
+export { prefixLayer } from './layers/prefixLayer';
+export { compose } from './utils/compose';
+export { Cache };
 
 export default Cache;
