+++ packages/cache-persist/src/layers/filterKeys.ts
@@ -1,11 +1,9 @@
 import { IMutateKey } from '../CacheTypes';
-import mutateKeys from './mutateKeys';
+import { mutateKeys } from './mutateKeys';
 
-function filterKeys(filter: (key: string) => boolean): IMutateKey {
+export function filterKeys(filter: (key: string) => boolean): IMutateKey {
     return mutateKeys(
         (key) => (filter(key) ? key : null),
         (key) => (filter(key) ? key : null),
     );
 }
-
-export default filterKeys;
