+++ packages/cache-persist/src/layers/jsonSerialize.ts
@@ -1,9 +1,7 @@
 import { IMutateValue } from '../CacheTypes';
-import mutateValues from './mutateValues';
+import { mutateValues } from './mutateValues';
 
-const jsonSerialize: IMutateValue = mutateValues(
+export const jsonSerialize: IMutateValue = mutateValues(
     (value) => JSON.stringify(value),
     (value) => JSON.parse(value),
 );
-
-export default jsonSerialize;
