+++ packages/cache-persist/src/layers/mutateKeys.ts
@@ -1,7 +1,5 @@
 import { IMutateKey } from '../CacheTypes';
 
-function mutateKeys(set: (key: string) => string | null, get: (key: string) => string | null): IMutateKey {
+export function mutateKeys(set: (key: string) => string | null, get: (key: string) => string | null): IMutateKey {
     return { set, get } as IMutateKey;
 }
-
-export default mutateKeys;
