+++ packages/cache-persist/src/layers/mutateValues.ts
@@ -1,7 +1,5 @@
 import { IMutateValue } from '../CacheTypes';
 
-function mutateValues(set: (value: any) => any, get: (value: any) => any): IMutateValue {
+export function mutateValues(set: (value: any) => any, get: (value: any) => any): IMutateValue {
     return { set, get } as IMutateValue;
 }
-
-export default mutateValues;
