+++ packages/cache-persist/src/utils/compose.ts
@@ -11,7 +11,7 @@
  *
  * REDUX
  */
-const compose = (...funcs) => {
+export const compose = (...funcs) => {
     if (funcs.length === 0) {
         return (arg) => arg;
     }
@@ -22,5 +22,3 @@ const compose = (...funcs) => {
 
     return funcs.reduce((a, b) => (...args) => a(b(...args)));
 };
-
-export default compose;
