--- packages/detect-network/src/deprecated/NetInfo.ts
@@ -1,4 +1,0 @@
-/* eslint import/no-extraneous-dependencies:0 */
-import { NetInfo } from 'react-native';
-
-export default NetInfo;
