--- packages/detect-network/src/deprecated/index.ts
@@ -1,3 +1,0 @@
-export { default as NetInfo } from './NetInfo';
-export { default as useNetInfo } from './useNetInfo';
-export { default as useIsConnected } from './useIsConnected';
