--- packages/detect-network/src/deprecated/useIsConnected.ts
@@ -1,8 +1,0 @@
-import NetInfo from './NetInfo';
-
-import useIsConnectedInternal from '../internal/useIsConnected';
-
-export function useIsConnected(): boolean {
-    return useIsConnectedInternal(NetInfo);
-}
-export default useIsConnected;
