--- packages/detect-network/src/deprecated/useNetInfo.ts
@@ -1,8 +1,0 @@
-import NetInfo from './NetInfo';
-
-import useNetInfoInternal from '../internal/useNetInfo';
-
-export function useNetInfo(): boolean {
-    return useNetInfoInternal(NetInfo);
-}
-export default useNetInfo;
