+++ packages/detect-network/src/useIsConnected.native.ts
@@ -1,9 +1,8 @@
 /* eslint import/no-extraneous-dependencies:0 */
-import NetInfo from '@react-native-community/netinfo';
+import { NetInfo } from '@wora/netinfo';
 
-import useIsConnectedInternal from './internal/useIsConnected';
+import { useIsConnected as useIsConnectedInternal } from './internal/useIsConnected';
 
 export function useIsConnected(): boolean {
     return useIsConnectedInternal(NetInfo);
 }
-export default useIsConnected;
