+++ packages/detect-network/src/useIsConnected.ts
@@ -1,7 +1,6 @@
 import { NetInfo } from '@wora/netinfo';
-import useIsConnectedInternal from './internal/useIsConnected';
+import { useIsConnected as useIsConnectedInternal } from './internal/useIsConnected';
 
 export function useIsConnected(): boolean {
     return useIsConnectedInternal(NetInfo);
 }
-export default useIsConnected;
