+++ packages/detect-network/src/useNetInfo.native.ts
@@ -1,4 +1,2 @@
 /* eslint import/no-extraneous-dependencies:0 */
-import { useNetInfo } from '@react-native-community/netinfo';
-
-export default useNetInfo;
+export { useNetInfo } from '@react-native-community/netinfo';
