+++ packages/redux/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@wora/redux",
-  "version": "2.0.0-rc.3",
+  "version": "2.0.0-rc.4",
   "keywords": [
     "cache",
     "wora",
