+++ packages/relay-store/src/RecordSource.ts
@@ -1,6 +1,6 @@
 import * as RelayRecordState from 'relay-runtime/lib/store/RelayRecordState';
 import { MutableRecordSource, Record, RecordMap } from 'relay-runtime/lib/store/RelayStoreTypes';
-import Cache, { ICache, DataCache, CacheOptions } from '@wora/cache-persist';
+import { Cache, ICache, DataCache, CacheOptions } from '@wora/cache-persist';
 
 const { EXISTENT, NONEXISTENT, UNKNOWN } = RelayRecordState;
